+++ ums-api/src/main/java/org/ums/manager/EquivalentCourseManager.java
@@ -4,6 +4,6 @@ import org.ums.domain.model.immutable.EquivalentCourse;
 import org.ums.domain.model.mutable.MutableEquivalentCourse;
 
 public interface EquivalentCourseManager extends
-    ContentManager<EquivalentCourse, MutableEquivalentCourse, Integer> {
+    ContentManager<EquivalentCourse, MutableEquivalentCourse, Long> {
 
 }
